in process...

+++ src/api/provider.js
@@ -1,3 +1,6 @@
+import Movie from '../models/movie.js';
+import nanoid from 'nanoid';
+
 export default class Provider {
   constructor(api, store) {
     this._api = api;
@@ -5,22 +8,93 @@ export default class Provider {
   }
 
   updateMovie(id, movie) {
-    this._api.updateMovie(id, movie);
+    const store = this._store.getAll();
+
+    if (this._isOnLine()) {
+      return this._api.updateMovie(id, movie)
+        .then((newMovie) => {
+          const rawMovie = {[newMovie.id]: newMovie.toRAW()};
+          this._store.setItem(`movies`, Object.assign({}, store.movies, rawMovie));
+
+          return newMovie;
+        });
+    }
+
+    const fakeUpdateMovie = Movie.parseMovie(Object.assign({}, movie.toRAW(), {id}));
+    store.movies[id] = Object.assign({}, fakeUpdateMovie.toRAW(), {offline: true});
+    this._store.setItem(`movies`, store.movies);
+
+    return Promise.resolve(fakeUpdateMovie);
   }
 
   getMovies() {
-    this._api.getMovies();
+    if (this._isOnLine()) {
+      return this._api.getMovies()
+        .then((movies) => {
+          const rawMovies = movies.map((movie) => movie.toRAW());
+          this._store.setItem(`movies`, Object.assign({}, rawMovies));
+
+          return movies;
+        });
+    }
+    const store = this._store.getAll();
+    const movies = Object.values(store.movies);
+
+    return Promise.resolve(Movie.parseMovies(movies));
   }
 
   getComments(movieId) {
-    this._api.getComments(movieId);
+    const store = this._store.getAll();
+    if (this._isOnLine()) {
+      return this._api.getComments(movieId)
+        .then((comments) => {
+          const movieCommments = {[movieId]: comments};
+          this._store.setItem(`comments`, Object.assign({}, store.comments, movieCommments));
+
+          return comments;
+        });
+    }
+
+    return Promise.resolve(store.comments[movieId]);
   }
 
   addComment(movieId, comment) {
-    this._api.addComment(movieId, comment);
+    const store = this._store.getAll();
+
+    if (this._isOnLine()) {
+      return this._api.addComment(movieId, comment)
+        .then((newComment) => {
+          store.comments[movieId] = Object.assign({}, newComment.comments);
+          this._store.setItem(`comments`, store.comments);
+
+          return newComment;
+        });
+    }
+    const fakeNewCommentId = nanoid();
+    const fakeNewComment = Object.assign({}, comment, {id: fakeNewCommentId, offline: true});
+
+    const comments = [...store.comments[movieId], fakeNewComment];
+    this._store.setItem(`comments`, Object.assign({}, store.comments, {[movieId]: comments}));
+
+    return Promise.resolve({movie: store.movies[movieId], comments});
   }
 
   deleteComment(id) {
-    this._api.deleteComment(id);
+    console.log('id: ', id);
+    if (this._isOnLine()) {
+      return this._api.deleteComment(id)
+        .then(() => {
+          this._store.removeItem(id);
+        });
+    }
+
+    const store = this._store.getAll();
+
+    console.log(store.comments);
+
+  }
+
+  _isOnLine() {
+    return window.navigator.onLine;
   }
 }
